Servlet Homework Chapter 37
In this assignment, I write an HTML form that prompts the user to enter loan amount, interest rate, and number of years. Clicking the Compute Loan Payment button should invoke a servlet to compute and display the monthly and total loan payments. The Loan class from Listing 10.2, Loan.java, computes the monthly and total payments.